Refactored entire website's style
