+++ public/index.html
@@ -1,34 +1,11 @@
 <!DOCTYPE html>
 <html>
-    <head>
-        <link rel="shortcut icon" href="favicon.ico" type="image/x-icon">
-        <meta charset="utf-8">
-        <title>Veggiebob's Webpage</title>
-    </head>
-    <body>
-        <!-- Circular floating link to GitHub Profile -->
-        <div id="root"></div>
-    </body>
-    <style>
-        .subpage-tile {
-            border-radius: 5px;
-            height: 150px;
-            width: 150px;
-        }
-        .img-link-wrapper {
-            display: inline-block;
-        }
-        .img-link {
-            border-radius: 5px;
-            padding: 30px;
-        }
-        #projects-box {
-            border-radius: 5px;
-            
-            /* center the contents */
-            display: flex;
-            justify-content: center;
-            align-items: center;
-        }
-    </style>
+  <head>
+    <link rel="shortcut icon" href="favicon.ico" type="image/x-icon" />
+    <meta charset="utf-8" />
+    <title>Veggiebob's Webpage</title>
+  </head>
+  <body>
+    <div id="root"></div>
+  </body>
 </html>
