+++ src/components/LoadingSpinner.jsx
@@ -1,0 +1,9 @@
+import React from "react";
+
+const LoadingSpinner = () => {
+  return (
+    <div className="border-2 rounded-full border-transparent border-t-orange-600 animate-spin w-12 h-12" />
+  );
+};
+
+export default LoadingSpinner;
