--- src/index.css
@@ -1,51 +1,0 @@
-@tailwind base;
-@tailwind components;
-@tailwind utilities;
-
-/* headers */
-
-h1 {
-  font-size: 2.5rem;
-  font-weight: 700;
-  line-height: 1.2;
-}
-
-h2 {
-  font-size: 2rem;
-  font-weight: 700;
-    line-height: 1.2;
-}
-
-h3 {
-  font-size: 1.75rem;
-  font-weight: 700;
-  line-height: 1.2;
-}
-
-h4 {
-  font-size: 1.5rem;
-  font-weight: 700;
-  line-height: 1.2;
-}
-
-/* other */
-
-input {
-  background-color: #ccc;
-  border-radius: 3px;
-  /* slight light shadow */
-  box-shadow: 0px 0px 3px #ccc;
-}
-
-a {
-  color: magenta;
-  font-weight: bold;
-  /* add underline */
-  text-decoration: underline;
-}
-
-body {
-  @apply m-3;
-  background-color: #333;
-  color: #ccc;
-}
