--- src/shadertoy.css
@@ -1,60 +1,0 @@
-.shadertoy-preview {
-    border-radius: 7px;
-    width: 80%;
-    /* slight shadow below */
-    box-shadow: -1px 5px 3px #222;
-}
-
-td p {
-    padding-bottom: 30px;
-}
-td img {
-    margin-bottom: 50px;
-}
-
-/* almost completely stolen from https://www.youtube.com/watch?v=wwJcYaJg8Mc */ 
-
-.spinner-wrapper{
-    top: 0;
-    left: 0;
-    z-index: 9999;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-}
-
-.spinner {
-    position: relative;
-    width: 4rem;
-    height: 4rem;
-    border-radius: 50%;
-}
-
-.spinner::before,
-.spinner:after{
-    content: "";
-    position: absolute;
-    border-radius: 50%;
-}
-
-.spinner:before {
-    width: 100%;
-    height: 100%;
-    background-image:linear-gradient(90deg, #ff00cc 0%,#333 100% );
-    animation: spin .7s infinite linear;
-}
-.spinner:after {
-    width: 90%;
-    height: 90%;
-    background-color: #333;
-    top: 50%;
-    left: 50%;
-    transform: translate(-50%, -50%);
-}
-
-@keyframes spin {
-    to {
-        transform: rotate(360deg);
-    }
-
-}
